refactor(offers): :bug: refactored data type

+++ lib/offers/domain/entities/salary.dart
@@ -1,7 +1,7 @@
 import 'package:teacher_finder/offers/domain/entities/enums/currency.dart';
 
 class Salary {
-  final int mount;
+  final double mount;
   final Currency currency;
 
   Salary({
